pyproject.toml
[project]
name = "jira"
version = "0.1.0"
description = "CLI tool for Jira using atlassian-python-api"
authors = []
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "atlassian-python-api>=3.41.4",
    "click>=8.1.7",  # For CLI interface
    "python-dotenv>=1.0.0",  # For environment variable management
    "tabulate>=0.9.0",  # For table formatting
]

[project.scripts]
jira = "jira_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jira_cli"] 